+++ src/services/api/paths/aws.ts
@@ -1,14 +1,21 @@
 import { publicClient } from "../clients";
 
 interface awsConnect {
-  otp: string;
+  workspaceId: string,
+  token: string;
   account: string;
   region: string;
   alias: string;
 }
 
 async function awsConnect(params: awsConnect): Promise<{ url: string }> {
-  const res = (await publicClient.put("/aws/connect", params)).data;
+  const res = (
+    await publicClient.put("/aws/connect", params, {
+      headers: {
+        authorization: `Bearer ${params.token}`,
+      },
+    })
+  ).data;
   return res;
 }